refactor: http 모듈 재사용 가능하도록 변경
