+++ app/src/main/java/woowacourse/shopping/data/server/ProductService.kt
@@ -1,0 +1,10 @@
+package woowacourse.shopping.data.server
+
+import woowacourse.shopping.domain.Product
+
+interface ProductService {
+    fun findAll(): List<Product>
+    fun find(id: Int): Product
+    fun findRange(mark: Int, rangeSize: Int): List<Product>
+    fun isExistByMark(mark: Int): Boolean
+}
