+++ app/src/main/java/woowacourse/shopping/view/productlist/ProductListActivity.kt
@@ -10,9 +10,9 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.GridLayoutManager
 import woowacourse.shopping.R
 import woowacourse.shopping.data.CartProductSqliteProductRepository
-import woowacourse.shopping.data.ProductMockWebRepository
 import woowacourse.shopping.data.RecentViewedDbRepository
 import woowacourse.shopping.data.db.CartDBHelper
+import woowacourse.shopping.data.server.ProductServiceImpl
 import woowacourse.shopping.databinding.ActivityProductListBinding
 import woowacourse.shopping.model.ProductModel
 import woowacourse.shopping.view.cart.CartActivity
@@ -44,7 +44,7 @@ class ProductListActivity : AppCompatActivity(), ProductListContract.View {
         presenter =
             ProductListPresenter(
                 this,
-                ProductMockWebRepository(),
+                ProductServiceImpl(),
                 RecentViewedDbRepository(this),
                 CartProductSqliteProductRepository(
                     CartDBHelper(this),
